Refactor "Universe repo moving" code.
